+++ sentry-kotlin-multiplatform/src/commonMain/kotlin/io/sentry/kotlin/multiplatform/SentryOptions.kt
@@ -70,6 +70,9 @@ public open class SentryOptions {
     /** Information about the Sentry SDK that generated this event. */
     public var sdk: SdkVersion? = null
 
+    /** Sets the minimum log level. Default is [SentryLevel.DEBUG] */
+    public var diagnosticLevel: SentryLevel = SentryLevel.DEBUG
+
     /** This variable controls the total amount of breadcrumbs that should be captured. Default is 100. */
     public var maxBreadcrumbs: Int = DEFAULT_MAX_BREADCRUMBS